rigid bodies now run pretty quickly
the slow part is now the derivative of the rotation matrix.

I had to make new interface to the potentials to make use of interaction
lists.  Each potential will have to be updated individually to take
advantage of the speed up.

The interaction list interface should be useful elsewhere besides rigid
bodies.

so far only the fortran lj potential is updated.