Add a version of mylbfgs which uses a callback function for the potential call
This is as close as we can get to using the GMIN lbfgs.

It uses roughly the same number of potential calls as other lbfgs
implementations, but it runs much faster for small systems.

The downside is that it runs as a black box.  We can't attach events to
be run every step unless we add another callback function, or piggyback
on the potential call.

Now that callback functions are working we can implement more involved
things from GMIN like hybrid lbfgs transition state optimization.