Change permutational optimization
This fixes what might be considered a bug in the minpermdist routines.
The cost matrix for optimizing permuations is now based on the distance
between atoms. Previously it was based on the distance squared between
atoms.  Having it based on distance squared emphasises many small
displacements over fewer longer displacements