Revert "Change permutational optimization"
This reverts commit 21442fbdfdb423ebe63debc0dd3ecfca27a292f6.

I think the cost matrix based on distance squared is actually correct.
The total mindist distance between two configurations is
np.linalg.norm(X1-X2) which is which is the root of the sum of the
squared distances between the atoms.  So we should do minimization on
the sum of the square distances.