add a new minpermdist routine
this new minpermdist routine iterates random rotations followed by
permuational alignment followed by rotational alignment in order to
stochastically find the best alignment.  It avoids using quenching of
the permutationally invariant potential.  This is more similar to what
is done in OPTIM.  It is several times faster and not significantly more
inaccurate.