Reorganize connect algorithm
move the local connect attempt (NEB + ts refinement) to a separate
class.  The new class LocalConnect will not have any knowledge of the
global graph or database.  It's cleaner this way and easier to
parallelize.

n.b. this breaks the parallel connect stuff.  That will be fixed in
future commits.