more modifications to new optimizers
noticed a "bug" in lbfgs_scipy.  The scipy version of lbfgs uses the
bounded L-BFGS-B algorithm.  The gradient is projected to avoid stepping
over the bounds, and the tolerance check is done on the projected
gradient.  I still don't know why, but this means that the actual
gradient never satisfies the required tolerance.  Very strange.