modify LBFGS and MYLBFGS so it can be safely restarted from new coords
I save the last step instead of the last coords.  That way the memory is
updated correctly before each step.  It would be cleaner to update the
memory at the end of each step, but MYLBFGS doesn't do it that way and
at the moment I want to keep LBFGS and MYLBFGS as similar as possible.