fix bug in atomic cluster pgorder calculation
During check_match() the rotational alignment should not be
re-optimized.  This was giving false matches, making the point group
order larger than it should be.

pgorder for the LJ6 and LJ75 global minimum are now computed correctly.

I've not yet fully checked that this does not cause any other problems