undefer coords in sqlalchemy query in addMinimum
This gives a massive speed up in the addMinimum function.

Minimum.coords (and other non-sql types) are set as deferred, which
means that they are not automatically loaded during a query.  The
following code

    for m in db.session.query(Minimum):
        print m.coords

will access the database once to get the list of mimima, then once for
each iteration to get the coords from the database.  The following code
gets everything at once and is much faster

    from sqlalchemy.orm import undefer
    for m in db.session.query(Minimum).options(undefer("coords")):
        print m.coords

Currently I've only changed it in Database.addMinimum, but we should
think about doing something similar (optionally) for Database.minima()