add commit_interval to Database.minimum_adder()
This makes it easy to reduce the frequency that db.session.commit() is
called.  This can make things run faster.  For example, if you set
commit_interval=100 it will only commit every 100 times addMinimum is
called.  Also, the function MinimumAdder.__del__() is overloaded so that
commit will always be called at the end of a run.