fix a bug in rigid body align
there was a typo in _cdef_MeasureAngleAxisCluster.align().  x1 was
passed as both x1 and x2.

Add tests for rigid body measure.align and transform.rotate.

Ensure exceptions are handled in some of the cdef functions.

Ensure findrotation returns a real object

Fix the documentation for rotations.rotate_aa.  The order of the inputs
is non-intuitive.